requirements.txt
torch>=1.5.0
mne>=0.20.3
pyyaml==5.3.1
pyyaml-include==1.2
numpy>=1.17.4
pandas>=1.0.0
tqdm>=4.46.0